Cancel Click/DoubleClick - Multibet
#1 Cancel Click/DoubleClick
Changed the logic.
From 2 stacks undoBets/tmpBets, we now use a undoBets Stack[8] array,
for each player.
Previous logic was slow when lots of players had bet and we tried to
cancel last bet or all the bets.
#2 Multibet
Changed the multibet to 5 credits per time instead of 1 and changed the
logic when either the player cannot bet 5 credits or the chip cannot
accept bets of 5 credits.